pyproject.toml
[project]
name = "ifvg-tempo-backtester"
version = "0.1.0"
description = "IFVG transcript extraction, rule versioning, and Python backtesting workflows."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "openai>=1.0.0",
  "pandas>=2.0.0",
  "python-dotenv>=1.0.0",
  "pydantic>=2.0.0",
  "pytz>=2024.0",
  "python-dateutil>=2.8.0",
  "pyyaml>=6.0.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
